+++ elasticsearch/CHANGELOG.md
@@ -1,9 +1,11 @@
+# 0.1.5
+[rajeshj] - Updating with multi tenant cluster configurations 
 # 0.1.4
 [rajeshj] - Adding application cluster configurations
 # 0.1.3
 [rajeshj] - Added elasticsearch node indices directory
 # 0.1.2
-[kiran] - Updated to create nginx conf.d directory if missing
+[kiran]   - Updated to create nginx conf.d directory if missing
 # 0.1.1
 [rajeshj] - Integrating with Nginx, and upgraded elasticsearch with 5.3.0
 # 0.1.0